Cygwin, MIB-II: Fix a few compiler warnings

+++ agent/mibgroup/mibII/data_access/at_iphlpapi.c
@@ -384,8 +384,8 @@ write_arp(int action,
              */
             if (!create_flag) {
                 if ((status = SetIpNetEntry(oldarp_row)) != NO_ERROR) {
-                    snmp_log(LOG_ERR, "Error in case UNDO, status : %lu\n",
-                             status);
+                    snmp_log(LOG_ERR, "Error in case UNDO, status : %u\n",
+                             (unsigned int)status);
                     retval = SNMP_ERR_UNDOFAILED;
                 }
             }
@@ -395,8 +395,8 @@ write_arp(int action,
 
                 if ((status = SetIpNetEntry(arp_row)) != NO_ERROR) {
                     snmp_log(LOG_ERR,
-                             "Error while deleting added row, status : %lu\n",
-                             status);
+                             "Error while deleting added row, status : %u\n",
+                             (unsigned int)status);
                     retval = SNMP_ERR_UNDOFAILED;
                 }
             }
@@ -415,8 +415,8 @@ write_arp(int action,
             if (arp_row->dwPhysAddrLen != 0) {
                 if ((status = CreateIpNetEntry(arp_row)) != NO_ERROR) {
                     snmp_log(LOG_ERR,
-                             "Inside COMMIT: CreateIpNetEntry failed, status %lu\n",
-                             status);
+                             "Inside COMMIT: CreateIpNetEntry failed, status %u\n",
+                             (unsigned int)status);
                     retval = SNMP_ERR_COMMITFAILED;
                 }
             } else {
